requirements.txt
Django,
Pillow,
psycopg[binary],
python-dotenv,
djangorestframework,